Encapsulated versioned read into service
